+++ src/cnpy++.cpp
@@ -398,7 +398,7 @@ std::vector<char>
 cnpypp::create_npy_header(cnpypp::span<uint64_t const> const shape,
                           cnpypp::span<std::string_view const> labels,
                           cnpypp::span<char const> dtypes,
-                          cnpypp::span<unsigned const> sizes,
+                          cnpypp::span<size_t const> sizes,
                           MemoryOrder memory_order) {
   std::vector<char> dict;
   append(dict, "{'descr': [");
